Move `Ref` to its own module
